# docs/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: rosmon HTTP API
  version: 1.0.0
  description: |
    REST + Server-Sent Events service di atas RouterOS API. Multi-router:
    semua endpoint (kecuali Device CRUD & Health) berada di bawah
    `/api/v1/devices/{device_id}/...` agar satu instance bisa melayani
    banyak router sekaligus.

    ## Response envelope

    Semua endpoint pakai envelope konsisten:

    - **Success single**: `{ "data": { ... } }`
    - **Success list**:   `{ "data": [ ... ], "meta": { "count": N } }`
    - **Error**:          `{ "error": { "code": "...", "message": "...", "path": "..." } }`

    Error `code` stabil — frontend bisa branch dengan aman. Mapping:

    | Code | HTTP | Trigger |
    |---|---|---|
    | `NOT_FOUND` | 404 | `mikrotik.ErrNotFound` |
    | `INVALID_ARGUMENT` | 400 | `mikrotik.ErrInvalidArgument` |
    | `AMBIGUOUS` | 409 | `mikrotik.ErrAmbiguous` |
    | `VALIDATION` | 400 | body gagal validasi `gin` binding |
    | `TIMEOUT` | 504 | `context.DeadlineExceeded` |
    | `CANCELED` | 499 | `context.Canceled` |
    | `INTERNAL` | 500 | uncategorized / panic recovery |
    | `DEVICE_NOT_FOUND` | 404 | device middleware: device ID tidak terhubung |
    | `SERVICE_UNAVAILABLE` | 503 | InfluxDB / dependency lain down |

    Setiap response menyertakan header `X-Request-ID` (16-char hex) untuk
    korelasi log.

    ## Authentication

    Service ini sekarang mendukung **JWT access + refresh token**
...

...
    4. `POST /auth/logout` revoke refresh token.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /devices/{device_id}/hotspot/profiles:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/CacheQuery'
      - $ref: '#/components/parameters/TtlQuery'
    get:
      tags:
        - Hotspot Profiles
      summary: List profiles
      description: List `/ip/hotspot/user/profile`. Cache opsional.
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HotspotProfileResponse'
        '500':
          $ref: '#/components/responses/ErrorInternal'
    post:
      tags:
        - Hotspot Profiles
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileCreateRequest'
            examples:
              vip_2m:
                summary: Profile VIP 2M/2M
                value:
                  name: vip-2m
                  address_pool: hs-pool
                  rate_limit: 2M/2M
                  shared_users: 1
                  on_login: ':log info "login $user from $address";'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
  /devices/{device_id}/hotspot/profiles/by-name/{name}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/NamePath'
    get:
      tags:
        - Hotspot Profiles
      summary: Get profile by name
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HotspotProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/hotspot/profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags:
        - Hotspot Profiles
      summary: Get profile by ID
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HotspotProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    put:
      tags:
        - Hotspot Profiles
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileUpdateRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags:
        - Hotspot Profiles
      summary: Delete profile (cascade)
      description: |
        Hapus profile. Kalau body `{name: ...}` di-set, workflow akan cleanup
        scheduler/script yang nama-nya match (cascade delete).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileDeleteRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /devices/{device_id}/ppp/profiles:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - PPP
      summary: List PPP profiles
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PPPProfileResponse'
    post:
      tags:
        - PPP
      summary: Create PPP profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PPPProfileCreateRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
  /devices/{device_id}/ppp/profiles/by-name/{name}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/NamePath'
    get:
      tags:
        - PPP
      summary: Get PPP profile by name
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PPPProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
  /devices/{device_id}/ppp/profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags:
        - PPP
      summary: Get PPP profile by ID
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PPPProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    put:
      tags:
        - PPP
      summary: Update PPP profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PPPProfileUpdateRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags:
        - PPP
      summary: Delete PPP profile
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (HS256). Diperoleh via `POST /auth/login`.
        Wajib di-include sebagai header `Authorization: Bearer <token>`
        kecuali untuk endpoint public (login, refresh, logout, /healthz).
  schemas:
    Envelope:
      type: object
      description: Response envelope generik. Field `data` dan `error` saling eksklusif.
      properties:
        data:
          description: Payload sukses. Tipe spesifik tergantung endpoint.
        meta:
          type: object
          additionalProperties: true
          description: Metadata opsional, biasanya berisi `count` untuk list.
          example:
            count: 47
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorEnvelope:
      type: object
      description: Envelope untuk response error (status 4xx/5xx).
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      description: Detail error yang stabil — frontend bisa branch via `code`.
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable error code untuk branching di client.
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - AMBIGUOUS
            - VALIDATION
            - TIMEOUT
            - CANCELED
            - INTERNAL
            - DEVICE_NOT_FOUND
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Pesan error human-readable.
          example: user with id *99 not found
        path:
          type: string
          description: Path request yang menghasilkan error.
          example: /api/v1/devices/rb-main/hotspot/users/*99
        details:
          description: Payload tambahan (mis. validator detail). Tipe bebas.
...
    IDResponse:
      type: object
      description: Body untuk endpoint create yang hanya kembalikan `.id`.
      required:
        - id
      properties:
        id:
          type: string
          description: RouterOS internal ID (`.id`).
          example: '*5'
...
    HotspotProfileResponse:
      type: object
      required:
        - id
        - name
        - shared_users
        - add_mac_cookie
        - transparent_proxy
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: default
        address_pool:
          type: string
          example: hs-pool
        rate_limit:
          type: string
          example: 2M/2M
        shared_users:
          type: integer
          minimum: 0
          example: 1
        status_autorefresh:
          type: string
          example: 1m
        on_login:
          type: string
          example: ':log info $user'
        on_logout:
          type: string
        parent_queue:
          type: string
        idle_timeout:
          type: string
          example: 5m
        keepalive_timeout:
          type: string
          example: 2m
        session_timeout:
          type: string
          example: 1h
        mac_cookie_timeout:
          type: string
          example: 3d
        add_mac_cookie:
          type: boolean
        transparent_proxy:
          type: boolean
    HotspotProfileCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: vip-2m
        address_pool:
          type: string
          example: hs-pool
        rate_limit:
          type: string
          example: 2M/2M
        shared_users:
          type: integer
          minimum: 0
          example: 1
        status_autorefresh:
          type: string
          example: 1m
        on_login:
          type: string
        parent_queue:
          type: string
    HotspotProfileUpdateRequest:
      type: object
      description: Sparse update.
      properties:
        name:
          type: string
        address_pool:
          type: string
        rate_limit:
          type: string
        shared_users:
          type: integer
          minimum: 0
          nullable: true
        status_autorefresh:
          type: string
        on_login:
          type: string
          nullable: true
        parent_queue:
          type: string
    HotspotProfileDeleteRequest:
      type: object
      description: |
        Optional body untuk DELETE profile. Kalau `name` di-set, workflow akan
        cleanup scheduler/script yang nama-nya match (cascade delete).
      properties:
        name:
          type: string
...
    PPPProfileResponse:
      type: object
      required:
        - id
        - name
        - disabled
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: default
        local_address:
          type: string
        remote_address:
          type: string
        rate_limit:
          type: string
          example: 10M/10M
          description: Format rx/tx.
        session_timeout:
          type: string
          example: 1h30m
          description: RouterOS duration.
        idle_timeout:
          type: string
          example: 10m
        parent_queue:
          type: string
          example: global
        on_up:
          type: string
          description: Script optional dijalankan saat session up.
        on_down:
          type: string
          description: Script optional dijalankan saat session down.
        disabled:
          type: boolean
        comment:
          type: string
    PPPProfileCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: vip
        local_address:
          type: string
        remote_address:
          type: string
        rate_limit:
          type: string
          example: 10M/10M
        session_timeout:
          type: string
          example: 1h30m
        idle_timeout:
          type: string
          example: 10m
        parent_queue:
          type: string
        on_up:
          type: string
        on_down:
          type: string
        disabled:
          type: boolean
          nullable: true
        comment:
          type: string
    PPPProfileUpdateRequest:
      type: object
      description: Sparse update — field yang tidak di-set tidak diubah.
      properties:
        name:
          type: string
        local_address:
          type: string
        remote_address:
          type: string
        rate_limit:
          type: string
        session_timeout:
          type: string
          nullable: true
        idle_timeout:
          type: string
          nullable: true
        parent_queue:
          type: string
          nullable: true
        on_up:
          type: string
          nullable: true
        on_down:
          type: string
          nullable: true
        disabled:
          type: boolean
          nullable: true
        comment:
          type: string
          nullable: true
...
  parameters:
    DeviceIdPath:
      name: device_id
      in: path
      required: true
      description: Numeric device ID.
      schema:
        type: integer
        format: int64
        example: 1
    IdPath:
      name: id
      in: path
      required: true
      description: RouterOS internal ID (`.id`), biasanya format `*N` (mis. `*1`, `*2A`).
      schema:
        type: string
        example: '*1'
    NamePath:
      name: name
      in: path
      required: true
      description: Nama resource (user/profile/secret) untuk lookup by-name.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        example: user01
...
    CacheQuery:
      name: cache
      in: query
      required: false
      description: Aktifkan caching untuk endpoint GET ini. Default `false`.
      schema:
        type: boolean
        default: false
    TtlQuery:
      name: ttl
      in: query
      required: false
      description: TTL cache dalam format Go duration (mis. `60s`, `5m`). Hanya berlaku saat `cache=true`.
      schema:
        type: string
        pattern: ^[0-9]+(ns|us|µs|ms|s|m|h)$
        example: 60s
...
  responses:
...
    ErrorValidation:
      description: Body request gagal validasi schema (gin binding).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validation_failed:
              summary: Field wajib tidak ada
              value:
                error:
                  code: VALIDATION
                  message: request validation failed
                  path: /api/v1/devices/rb-main/hotspot/users
                  details: 'Key: ''HotspotUserCreateRequest.Name'' Error:Field validation for ''Name'' failed on the ''required'' tag'
    ErrorNotFound:
      description: Resource yang dirujuk tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            user_not_found:
              summary: User tidak ada
              value:
                error:
                  code: NOT_FOUND
                  message: 'mikrotik: not found'
                  path: /api/v1/devices/rb-main/hotspot/users/*99
...
    ErrorInternal:
      description: Error internal server. Dipakai untuk uncategorized + panic recovery.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internal:
              summary: Unexpected error
              value:
                error:
                  code: INTERNAL
                  message: unexpected error during operation
                  path: /api/v1/devices/rb-main/hotspot/users
...
    NoContentResponse:
      description: Operasi sukses tanpa body response (204 No Content).